moved image processing to separate util
