+++ cvat-ui/src/utils/opencv-wrapper/histogram-equalization.ts
@@ -2,10 +2,10 @@
 //
 // SPDX-License-Identifier: MIT
 
-import { ImageProcessing } from './opencv-interfaces';
+import { ImageProcessing } from 'utils/image-processing';
 
 export interface HistogramEqualization extends ImageProcessing{
-    processImage: (src:ImageData, frameNumber: number)=>ImageData;
+    processImage: (src:ImageData, frameNumber: number) => ImageData;
 }
 
 interface HashedImage{
